refactor: 모달 기능 별 컴포넌트로 분리, timer 객체를 timerObject로 정의
