+++ src/atom.js
@@ -12,8 +12,23 @@ const localStorageEffect =
 			isReset ? localStorage.removeItem(key) : localStorage.setItem(key, JSON.stringify(newValue))
 		})
 	}
+export const timerObject = (time, name, isRunning = false, disabled = false) => {
+	return { time, name, isRunning, disabled }
+}
+export const [defaultMode, addMode, doneMode] = ['defaultMode', 'addMode', 'doneMode']
+
 export const timerState = atom({
 	key: 'timer',
-	default: {}, //{id:{time, name, isRunning}}, anotherId: {time,name, isRunning}}
+	default: {}, //timerObject
 	effects: [localStorageEffect('timer')],
+})
+
+export const modeState = atom({
+	key: 'mode',
+	default: defaultMode,
+})
+
+export const originIdState = atom({
+	key: 'originId',
+	default: null,
 })
