+++ src/pages/updateTimeDivider.jsx
@@ -1,20 +1,17 @@
-import React, { useState } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import NavBar from '../components/NavBar'
-import { useRecoilState } from 'recoil'
-import { timerState } from '../atom'
+import { useRecoilState, useSetRecoilState } from 'recoil'
+import { addMode, defaultMode, doneMode, modeState, originIdState, timerState } from '../atom'
 import Timer from '../components/Timer'
 import Button from '../components/Button'
-import Select from '../components/Select'
-import { HOUR_NUMBERS, MINUTE_NUMBERS } from '../components/TimeSelectForm'
-import FormModal from '../components/FormModal'
+import AddFormModal from '../components/AddFormModal'
+import DoneFormModal from '../components/DoneFormModal'
 
 const UpdateTimeDivider = () => {
-	const [updateMode, addMode, doneMode] = ['updateMode', 'addMode', 'doneMode']
-
 	const [timers, setTimers] = useRecoilState(timerState)
-	const [mode, setMode] = useState(updateMode)
-	const [originId, setOriginId] = useState(null)
+	const [mode, setMode] = useRecoilState(modeState)
+	const setOriginId = useSetRecoilState(originIdState)
 
 	const toggleTimerRunning = (id = '') => {
 		const newTimers = Object.assign({}, timers)
@@ -26,46 +23,7 @@ const UpdateTimeDivider = () => {
 		}
 		setTimers(newTimers)
 	}
-	const deleteTimer = id => {
-		const newTimers = Object.assign({}, timers)
-		delete newTimers[id]
-		setTimers(newTimers)
-	}
-	const addTimer = (name, time, id) => {
-		setTimers({ ...timers, [id]: { time, name } })
-	}
-	const mergeTimer = (originId, targetId) => {
-		const newTimers = {
-			...timers,
-			[targetId]: {
-				...timers[targetId],
-				time: timers[targetId].time + timers[originId].time,
-			},
-		}
-		delete newTimers[originId]
-		setOriginId(null)
-		setTimers(newTimers)
-	}
 
-	const onAddEvent = e => {
-		e.preventDefault()
-		const [name, time, id] = [
-			e.target.name.value,
-			hmsToTime(e.target.hour.value, e.target.minute.value),
-			'' + Date.now(),
-		]
-		addTimer(name, time, id)
-	}
-	const onMergeEvent = e => {
-		e.preventDefault()
-		if (!originId) return
-		mergeTimer(originId, e.target.targetId.value)
-	}
-	const onDeleteEvent = e => {
-		if (!originId) return
-		deleteTimer(originId)
-		setOriginId(null)
-	}
 	return (
 		<Wrapper>
 			<NavBar>모래시계 편집하기</NavBar>
@@ -83,72 +41,28 @@ const UpdateTimeDivider = () => {
 					size={'md'}
 					onClick={() => {
 						toggleTimerRunning()
-						mode === doneMode ? setMode(updateMode) : setMode(doneMode)
+						mode === doneMode ? setMode(defaultMode) : setMode(doneMode)
 					}}
 				>
 					{mode === doneMode ? '취소' : '완료하기'}
 				</Button>
 			</ToolBar>
 			<TimerArea mode={mode}>
-				{Object.entries(timers).map(([id, { time, name }], index) => (
+				{Object.entries(timers).map(([id, { time, name, disabled }], index) => (
 					<Timer
 						key={id}
 						id={id}
 						name={name}
 						expiryTimestamp={timeToExpiryTime(time)}
 						onClick={() => {
+							if (timers[id].disabled) return
 							mode === doneMode ? setOriginId(id) : toggleTimerRunning(id)
 						}}
-						onExpire={() => deleteTimer(id)}
 					/>
 				))}
 			</TimerArea>
-			<FormModal
-				id={'addForm'}
-				visible={mode === addMode}
-				onClose={() => setMode(updateMode)}
-				onSubmit={e => {
-					onAddEvent(e)
-					setMode(updateMode)
-				}}
-				onCancel={e => {
-					setMode(updateMode)
-				}}
-				titleText={'추가하기'}
-				cancelText={'취소'}
-				confirmText={'확인'}
-			>
-				<>
-					할 일: <input name={'name'} required={true} type={'text'} maxLength={10} />
-					몇 시간 : <Select name={'hour'} data={HOUR_NUMBERS} />
-					몇 분: <Select name={'minute'} data={MINUTE_NUMBERS} />
-				</>
-			</FormModal>
-			<FormModal
-				id={'doneForm'}
-				visible={mode === doneMode && originId}
-				onClose={() => setOriginId(null)}
-				onSubmit={e => onMergeEvent(e)}
-				onCancel={e => onDeleteEvent(e)}
-				titleText={'남은 시간을 어느 항목에 합치시겠습니까?'}
-				cancelText={'시간버리기'}
-				confirmText={'합치기'}
-			>
-				<>
-					{originId && timers[originId].name}
-					에서
-					<select name={'targetId'}>
-						{Object.entries(timers).map(
-							([optionId, { time, name }]) =>
-								optionId !== originId && (
-									<option key={optionId} value={optionId}>
-										{name}
-									</option>
-								),
-						)}
-					</select>
-				</>
-			</FormModal>
+			<AddFormModal />
+			<DoneFormModal />
 		</Wrapper>
 	)
 }
@@ -158,9 +72,7 @@ export default UpdateTimeDivider
 const timeToExpiryTime = time => {
 	return new Date(new Date().getTime() + time * 1000)
 }
-const hmsToTime = (hour = 0, minute = 0, seconds = 0) => {
-	return hour * 60 * 60 + minute * 60 + seconds
-}
+
 const Wrapper = styled.div`
 	display: flex;
 	flex-direction: column;
